Fix valid_pieces_for_player
The issue was that the list would be pieces from the sim board which have already moved. The list is now filled with piece
from the base board rather than the sim board. This means the list now represent the correct pieces that the player can
move without causing a check onto themselves.